table_cropper/pyproject.toml
[project]
name = "table-cropper"
version = "0.1.0"
description = "Crop tables (with titles and captions) from scientific journal PDFs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "timm>=0.9.0",
    "requests>=2.28.0",
    "scipy>=1.10.0",
    "ultralytics>=8.3.0",
    "datasets>=2.14.0",
    "python-dotenv>=1.2.2",
]
